:memo: docs, move code in in processor/__init__.py to processor/base.py
