+++ ocrd/webservice/processor.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask import request
 
-from ocrd.processor import run_processor
+from ocrd.processor.base import run_processor
 from ocrd.processor.characterize.exif import ExifProcessor
 from ocrd.processor.segment_line.tesseract3 import Tesseract3LineSegmenter
 from ocrd.processor.segment_region.tesseract3 import Tesseract3RegionSegmenter
